chore: unify kill local process into a process mod
