+++ crates/fluvio-cluster/src/cli/shutdown.rs
@@ -1,16 +1,14 @@
-use std::ffi::OsString;
-use std::fs::remove_file;
 use std::process::Command;
 
 use anyhow::bail;
 use anyhow::Result;
 use clap::Parser;
 use tracing::debug;
-use sysinfo::System;
 
 use fluvio_types::defaults::SPU_MONITORING_UNIX_SOCKET;
 use fluvio_command::CommandExt;
 
+use crate::process;
 use crate::render::ProgressRenderer;
 use crate::cli::ClusterCliError;
 use crate::progress::ProgressBarFactory;
@@ -36,7 +34,7 @@ impl ShutdownOpt {
 
         match installation_type {
             InstallationType::Local | InstallationType::LocalK8 | InstallationType::ReadOnly => {
-                Self::kill_local_processes(&installation_type, &pb).await?;
+                Self::shutdown_local(&installation_type, &pb).await?;
             }
             InstallationType::Cloud => {
                 let profile = config.config().current_profile_name().unwrap_or("none");
@@ -50,67 +48,23 @@ impl ShutdownOpt {
         Ok(())
     }
 
-    async fn kill_local_processes(
+    async fn shutdown_local(
         installation_type: &InstallationType,
         pb: &ProgressRenderer,
     ) -> Result<()> {
-        pb.set_message("Uninstalling fluvio local components");
-
-        let kill_proc = |name: &str, command_args: Option<&[String]>| {
-            sysinfo::set_open_files_limit(0);
-            let mut sys = System::new();
-            sys.refresh_processes(sysinfo::ProcessesToUpdate::All); // Only load what we need.
-            for process in sys.processes_by_exact_name(name.as_ref()) {
-                if let Some(cmd_args) = command_args {
-                    let proc_cmds = process.cmd();
-                    if cmd_args.len() > proc_cmds.len() {
-                        continue; // Ignore procs with less command_args than the target.
-                    }
-                    if cmd_args
-                        .iter()
-                        .map(OsString::from)
-                        .collect::<Vec<_>>()
-                        .iter()
-                        .ne(proc_cmds[..cmd_args.len()].iter())
-                    {
-                        continue; // Ignore procs which don't match.
-                    }
-                }
-                if !process.kill() {
-                    // This will fail if called on a proc running as root, so only log failure.
-                    debug!(
-                        "Sysinto process.kill() returned false. pid: {}, name: {}: user: {:?}",
-                        process.pid(),
-                        process.name().to_str().unwrap_or("unknown"),
-                        process.user_id(),
-                    );
-                }
-            }
-        };
-        kill_proc("fluvio", Some(&["cluster".into(), "run".into()]));
-        kill_proc("fluvio", Some(&["run".into()]));
-        kill_proc("fluvio-run", None);
+        process::kill_local_processes(pb).await?;
 
         if let InstallationType::LocalK8 = installation_type {
             let _ = Self::remove_custom_objects("spus", true);
         }
 
         // remove monitoring socket
-        match remove_file(SPU_MONITORING_UNIX_SOCKET) {
-            Ok(_) => {
-                pb.println(format!(
-                    "Removed spu monitoring socket: {SPU_MONITORING_UNIX_SOCKET}"
-                ));
-            }
-            Err(io_err) if io_err.kind() == std::io::ErrorKind::NotFound => {
-                debug!("SPU monitoring socket not found: {SPU_MONITORING_UNIX_SOCKET}");
-            }
-            Err(err) => {
-                pb.println(format!(
-                    "SPU monitoring socket  {SPU_MONITORING_UNIX_SOCKET}, can't be removed: {err}"
-                ));
-            }
-        }
+        process::delete_fs(
+            Some(SPU_MONITORING_UNIX_SOCKET),
+            "SPU monitoring socket",
+            true,
+            Some(pb),
+        );
 
         pb.println("Uninstalled fluvio local components");
         pb.finish_and_clear();
