+++ crates/fluvio-cluster/src/lib.rs
@@ -34,6 +34,7 @@ mod delete;
 mod error;
 mod progress;
 pub mod runtime;
+mod process;
 
 /// extensions
 #[cfg(feature = "cli")]
